# compare performance of source terms in 3D