--- a/src/equations/1d/compressible_euler.jl
+++ b/src/equations/1d/compressible_euler.jl
@@ -151,8 +151,7 @@ end
 
 
 # Apply source terms
-# TODO: Taal remove methods with the signature below?
-#       Or keep them as an option for possiby increased performance?
+# TODO: Taal remove methods with the signature below
 function source_terms_convergence_test(ut, u, x, element_id, t, n_nodes, equation::CompressibleEulerEquations1D)
   # Same settings as in `initial_condition`
   c = 2