… examples/3d/elixir_euler_mortar.jl
… examples/3d/elixir_eulergravity_eoc_test.jl
… examples/3d/elixir_hyp_diff_nonperiodic.jl
… examples/3d/parameters_eulergravity_eoc_test.toml
… examples/paper-self-gravitating-gas-dynamics/parameters_eulergravity_eoc_test.toml
… test/test_examples_3d.jl
… test/test_examples_3d_old.jl
… test/test_paper-self-gravitating-gas-dynamics.jl