I used the following code to set up benchmarks (adapted to the examples).
```bash
julia --threads=1 --check-bounds=no
```
#### Taam

```julia
julia> using BenchmarkTools, Revise, Cthulhu, Test; using Trixi

julia> Trixi.init_parameters("examples/3d/parameters_euler_mortar.toml")

julia> mesh, solver, time_parameters, time_integration_function = Trixi.init_simulation();

julia> Trixi.run_simulation(mesh, solver, time_parameters, time_integration_function)

julia> @benchmark Trixi.calc_sources!(
           $(solver),
           $(solver.source_terms),
           $(Trixi.parameter("t_end")))
```
#### Taal
```julia
julia> using BenchmarkTools, Revise, Cthulhu, Test; using Trixi

julia> trixi_include("examples/3d/elixir_euler_mortar.jl")

julia> @benchmark Trixi.calc_sources!(
           $(Trixi.wrap_array(copy(sol.u[end]), semi)),
           $(Trixi.wrap_array(copy(sol.u[end]), semi)),
           $(sol.t[end]),
           $(semi.source_terms),
           $(equations),
           $(solver),
           $(semi.cache))
```


### `examples/3d/elixir_euler_source_terms.jl`

#### Taam
```
Section                             ncalls     time   %tot     avg     alloc   %tot      avg
--------------------------------------------------------------------------------------------
     source terms                     1.54k    116ms  18.3%  74.9μs   1.25MiB  4.55%     848B

BenchmarkTools.Trial:
  memory estimate:  848 bytes
  allocs estimate:  6
  --------------
  minimum time:     75.516 μs (0.00% GC)
  median time:      76.715 μs (0.00% GC)
  mean time:        80.521 μs (0.00% GC)
  maximum time:     280.836 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1
```

#### Taal
```
Section                ncalls     time   %tot     avg     alloc   %tot      avg
-------------------------------------------------------------------------------
  source terms            951   79.6ms  21.9%  83.7μs    966KiB  7.91%  1.02KiB

BenchmarkTools.Trial:
  memory estimate:  928 bytes
  allocs estimate:  6
  --------------
  minimum time:     85.270 μs (0.00% GC)
  median time:      87.191 μs (0.00% GC)
  mean time:        89.418 μs (0.00% GC)
  maximum time:     202.735 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1
```

### `examples/3d/elixir_euler_mortar.jl`

#### Taam
```
Section                             ncalls     time   %tot     avg     alloc   %tot      avg
--------------------------------------------------------------------------------------------
    source terms                       620   89.7ms  16.9%   145μs    513KiB  3.42%     848B

BenchmarkTools.Trial:
  memory estimate:  848 bytes
  allocs estimate:  6
  --------------
  minimum time:     136.219 μs (0.00% GC)
  median time:      138.133 μs (0.00% GC)
  mean time:        144.758 μs (0.00% GC)
  maximum time:     370.309 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1
```

#### Taal
```
Section                ncalls     time   %tot     avg     alloc   %tot      avg
-------------------------------------------------------------------------------
  source terms            636    105ms  20.2%   165μs    646KiB  5.40%  1.02KiB

BenchmarkTools.Trial:
  memory estimate:  928 bytes
  allocs estimate:  6
  --------------
  minimum time:     154.726 μs (0.00% GC)
  median time:      157.977 μs (0.00% GC)
  mean time:        161.632 μs (0.00% GC)
  maximum time:     318.420 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1
```

### `examples/3d/elixir_hyp_diff_llf.jl`

#### Taam
I used `time_integration_scheme = timestep_carpenter_kennedy_erk54_2N!`
with `cfl = 0.4` since the optimized 3S* method isn't ported yet.
```
Section                             ncalls     time   %tot     avg     alloc   %tot      avg
--------------------------------------------------------------------------------------------
    source terms                     7.14k    7.37s  47.8%  1.03ms   5.77MiB  2.55%     848B

BenchmarkTools.Trial:
  memory estimate:  848 bytes
  allocs estimate:  6
  --------------
  minimum time:     983.161 μs (0.00% GC)
  median time:      993.635 μs (0.00% GC)
  mean time:        1.017 ms (0.00% GC)
  maximum time:     1.517 ms (0.00% GC)
  --------------
  samples:          4916
  evals/sample:     1
```

#### Taal
```
Section                ncalls     time   %tot     avg     alloc   %tot      avg
-------------------------------------------------------------------------------
  source terms          3.57k    3.72s  49.7%  1.04ms   3.76MiB  3.08%  1.08KiB

BenchmarkTools.Trial:
  memory estimate:  960 bytes
  allocs estimate:  6
  --------------
  minimum time:     968.202 μs (0.00% GC)
  median time:      981.306 μs (0.00% GC)
  mean time:        1.004 ms (0.00% GC)
  maximum time:     1.873 ms (0.00% GC)
  --------------
  samples:          4981
  evals/sample:     1
```

### `examples/3d/elixir_hyp_diff_nonperiodic.jl`

#### Taam
I used `time_integration_scheme = timestep_carpenter_kennedy_erk54_2N!`
with `cfl = 0.4` since the optimized 3S* method isn't ported yet.
```
Section                             ncalls     time   %tot     avg     alloc   %tot      avg
--------------------------------------------------------------------------------------------
    source terms                     4.48k    1.23s  41.8%   275μs   3.62MiB  4.28%     848B

BenchmarkTools.Trial:
  memory estimate:  848 bytes
  allocs estimate:  6
  --------------
  minimum time:     262.515 μs (0.00% GC)
  median time:      266.568 μs (0.00% GC)
  mean time:        273.177 μs (0.00% GC)
  maximum time:     498.764 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1
```

#### Taal
```
Section                ncalls     time   %tot     avg     alloc   %tot      avg
-------------------------------------------------------------------------------
  source terms          4.48k    1.05s  39.0%   234μs   4.72MiB  4.97%  1.08KiB

BenchmarkTools.Trial:
  memory estimate:  960 bytes
  allocs estimate:  6
  --------------
  minimum time:     197.238 μs (0.00% GC)
  median time:      200.645 μs (0.00% GC)
  mean time:        208.289 μs (0.00% GC)
  maximum time:     442.517 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1
```

### `examples/3d/elixir_eulergravity_eoc_test.jl`
Here, I benchmarked the source term calculation of the Euler solver in more detail.
#### Taam
```
Section                                 ncalls     time   %tot     avg     alloc   %tot      avg
------------------------------------------------------------------------------------------------
    gravity solver                         110    8.99s  98.5%  81.8ms    670MiB  98.3%  6.09MiB
        source terms                     49.8k    547ms  5.99%  11.0μs   40.2MiB  5.91%     848B
    Euler solver                           110   39.8ms  0.44%   362μs   1.40MiB  0.21%  13.0KiB
      source terms                         110   9.13ms  0.10%  83.0μs   91.1KiB  0.01%     848B

julia> @benchmark Trixi.calc_sources!(
           $(solver[1]),
           $(solver[1].source_terms),
           $(Trixi.parameter("t_end")))
BenchmarkTools.Trial:
  memory estimate:  848 bytes
  allocs estimate:  6
  --------------
  minimum time:     80.742 μs (0.00% GC)
  median time:      82.682 μs (0.00% GC)
  mean time:        84.673 μs (0.00% GC)
  maximum time:     200.525 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1
```

#### Taal
```
Section                  ncalls     time   %tot     avg     alloc   %tot      avg
---------------------------------------------------------------------------------
gravity solver              121    8.96s  97.9%  74.1ms    548MiB  97.5%  4.52MiB
    source terms          52.9k    511ms  5.59%  9.67μs   55.7MiB  9.92%  1.08KiB
Euler solver                121   43.6ms  0.48%   360μs   1.14MiB  0.20%  9.67KiB
    source terms            121   11.8ms  0.13%  97.5μs    123KiB  0.02%  1.02KiB

julia> @benchmark Trixi.calc_sources!(
           $(Trixi.wrap_array(copy(sol.u[end]), semi)),
           $(Trixi.wrap_array(copy(sol.u[end]), semi)),
           $(sol.t[end]),
           $(semi.semi_euler.source_terms),
           $(equations_euler),
           $(solver_euler),
           $(semi_euler.cache))
BenchmarkTools.Trial:
  memory estimate:  928 bytes
  allocs estimate:  6
  --------------
  minimum time:     84.501 μs (0.00% GC)
  median time:      86.401 μs (0.00% GC)
  mean time:        88.911 μs (0.00% GC)
  maximum time:     244.091 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1
```

## Conclusions

There can be a slight performance impact of using point-wise source terms in Taal vs. element-wise source terms in Taam.
However, this performance impact is often smaller than the impact of optimizing the code itself, e.g. by using
`sinpi(...)` instead of `sin(pi * ...)` or `sincos(...)` instead of `sin(...), cos(...)`. Moreover, a performance
impact occurs mostly for the Euler equations, where it is not really important. In contrast, there is no negative
performance impact on the hyperbolic diffusion system, where the source terms are more important. Hence, I think it's
fine to keep point-wise source terms in Taal to get a clear separation of the solver and the equations.

Closes #208