@sloede 2020-11-02
> Hence, I think it's fine to keep point-wise source terms in Taal to get a clear separation of the solver and the equations.

I agree (will do a review shortly). However, can you explain the increased memory footprint? By my account, most source terms should be completely allocation-free, unless there's type instability from the `equations` etc. Any ideas?