@ranocha 2020-11-02
> However, can you explain the increased memory footprint? By my account, most source terms should be completely allocation-free, unless there's type instability from the `equations` etc. Any ideas?

We're using threads in the for loops, which create closures. Their size depends on the arguments being used - that explains this difference. There are no allocations if I remove `Threads.@threads` in `calc_sources!`.