@sloede 2020-11-02
> > However, can you explain the increased memory footprint? By my account, most source terms should be completely allocation-free, unless there's type instability from the `equations` etc. Any ideas?
> 
> We're using threads in the for loops, which create closures. Their size depends on the arguments being used - that explains this difference. There are no allocations if I remove `Threads.@threads` in `calc_sources!`.

Ah, yes, we had this before. Say, is this something we should file an issue for with julialang? I guess they are probably aware of this, but it might help to show that there are real-world codes out there who suffer from it.