@ranocha 2020-11-02
> Say, is this something we should file an issue for with julialang? I guess they are probably aware of this, but it might help to show that there are real-world codes out there who suffer from it.

You could do that, but I'm not sure whether you will get positive reactions.