@ranocha 2020-11-02
The failing test for `examples/2d/elixir_mhd_rotor_shockcapturing_amr.jl` seems to be the same kind of problem as in https://github.com/trixi-framework/Trixi.jl/pull/275#issuecomment-720352787.